Implementing templated connectivity

+++ utils/Options.h
@@ -369,7 +369,7 @@ class BoolOption : public Option
         fprintf(stderr, "  -%s, -no-%s", name, name);
 
         //Note: Changed this to signed int, otherwise this can overflow if name is longer than 32...
-        for (int i = 0; i < 32 - strlen(name)*2; i++)
+        for (int i = 0; i < 32 - (int)( strlen(name))*2; i++)
             fprintf(stderr, " ");
